pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azure-voice-language-lab"
version = "1.1.0"
description = "Laboratório completo para exploração de Azure Speech e Language Studio"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Rone Bragaglia", email = "ronbragaglia@gmail.com"}
]
maintainers = [
    {name = "Rone Bragaglia", email = "ronbragaglia@gmail.com"}
]
keywords = [
    "azure",
    "speech",
    "language",
    "nlp",
    "tts",
    "stt",
    "translation",
    "sentiment-analysis",
    "cognitive-services",
    "machine-learning",
    "ai",
    "voice",
    "text-to-speech",
    "speech-to-text"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Text Processing :: Linguistic",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Natural Language :: Portuguese (Brazilian)",
    "Natural Language :: English",
]
requires-python = ">=3.8"
dependencies = [
    "azure-cognitiveservices-speech>=1.46.0",
    "azure-ai-textanalytics>=5.3.0",
    "azure-ai-translation-text>=1.0.0",
    "azure-ai-language>=1.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.6.0",
]

[project.optional-dependencies]
web = [
    "streamlit>=1.31.0",
    "gradio>=4.20.0",
]
audio = [
    "pydub>=0.25.0",
    "soundfile>=0.12.0",
]
visualization = [
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "isort>=5.13.0",
    "pre-commit>=3.6.0",
]
docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autodoc-typehints>=2.0.0",
]
all = [
    "azure-voice-language-lab[web,audio,visualization,dev,docs]",
]

[project.urls]
Homepage = "https://github.com/Ronbragaglia/Azure-Voice-Language-Lab"
Documentation = "https://github.com/Ronbragaglia/Azure-Voice-Language-Lab#readme"
Repository = "https://github.com/Ronbragaglia/Azure-Voice-Language-Lab"
Issues = "https://github.com/Ronbragaglia/Azure-Voice-Language-Lab/issues"
Changelog = "https://github.com/Ronbragaglia/Azure-Voice-Language-Lab/blob/main/CHANGELOG.md"

[project.scripts]
azure-speech-lab = "src.cli:main"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*", "examples*", "docs*", "web*"]

[tool.setuptools.package-data]
"src" = ["*.json", "*.yaml", "*.yml"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | env
  | venv
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
skip = [".git", ".venv", "venv", "env", "build", "dist"]

# mypy configuration
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_optional = true
show_error_codes = true
show_column_numbers = true
show_error_context = true
color_output = true
error_summary = true
exclude = [
    "venv",
    ".venv",
    "env",
    "build",
    "dist",
    "tests",
]

[[tool.mypy.overrides]]
module = [
    "azure.*",
    "streamlit.*",
    "gradio.*",
    "pydub.*",
]
ignore_missing_imports = true

# pytest configuration
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "speech: marks tests related to speech recognition/synthesis",
    "language: marks tests related to language analysis",
    "translation: marks tests related to translation",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
    "*/cli.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

# flake8 configuration
[tool.flake8]
max-line-length = 100
extend-ignore = [
    "E203",  # whitespace before ':'
    "W503",  # line break before binary operator
    "E501",  # line too long (handled by black)
]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "env",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
]

# bandit configuration (security linter)
[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101", "B601"]
